requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.13
langchain-core==1.2.22
langchain-openai==1.1.12
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.18.1
numpy==2.4.3
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
pillow==11.3.0
protobuf==5.29.6
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.41.1
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
